+++ CHANGELOG.md
@@ -5,6 +5,13 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Changed
+- Renamed workflow property `currentEvent` to `event`
+- Renamed workflow function `transition` to `trigger`
+
+### Added
+- Alternative HTTP route to execute commands
+- New property `latestEventPosition` inside the context of commands
 
 ## [1.1.1] - 2021-04-19
 ### Changed